requirements.txt
alembic==1.5.2
appdirs==1.4.3
astroid==2.4.1
Babel==2.9.0
beautifulsoup4==4.9.0
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==5.0.1
crayons==0.4.0
cssselect==1.1.0
cycler==0.10.0
distlib==0.3.1
fake-useragent==0.1.11
filelock==3.0.12
Flask==1.0.3
Flask-Migrate==2.6.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
html5lib==1.1
httplib2==0.18.1
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.2
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.5
parse==1.15.0
protobuf==3.12.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==7.0.1
pylint==2.5.2
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.1
python-dateutil==2.6.0
python-decouple==3.4
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.1
requests==2.23.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.22
toml==0.10.0
tqdm==4.46.0
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.4.2
w3lib==1.21.0
webdriver-manager==3.3.0
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
XlsxWriter==1.2.9
